refactor: organize components into folders
