+++ src/app/viewer/viewer.component.css
