+++ src/app/viewer/viewer.component.spec.ts
