+++ src/app/viewer/viewer.component.ts
